Split a very long file into several
